[ADD] Different page-tree heights for different subtrees (UberPage => RevisionRoot, RevisionRoot => NodePages, RevisionRoot/ValuePage => NodePages, RevisionRoot/PathSummaryPage). [MOD] API changes, now the revisionNumber is an integer which should be more than sufficient. [REM] EStoragePaths, was replaced by DatabaseConfiguration.Paths. [ADD] Resource String/ID mapping in the session for upcoming brackit(.org) binding.

+++ bundles/sirix-core/src/main/java/org/sirix/access/AbsForwardingNodeReadTrx.java
@@ -59,7 +59,7 @@ public abstract class AbsForwardingNodeReadTrx extends ForwardingObject
   }
 
   @Override
-  public long getRevisionNumber() throws SirixIOException {
+  public int getRevisionNumber() throws SirixIOException {
     return delegate().getRevisionNumber();
   }
